Module_B/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.8.0
python-multipart>=0.0.9
mysql-connector-python>=9.0.0
aiohttp>=3.10.0
xxhash>=3.0.0
